Added api endpoints with mock data, backend side is ready
Created a db.json file in the root directory, this is what the json-server command uses for simulating the backend api.

To start the api type the following command: `npx json-server --watch db.json`

There are two endpoints currently, the /statistics endpoint holds the resources for all of the example diagrams.

TD: #12, #13